+++ CHANGELOG.md
@@ -54,6 +54,9 @@
     - Add Dataset chunked() factory for online training
     - Added L1 penalty to `Dense` layers
     - `Adaline`, `Logistic Regression`, `Softmax Classifier` now elastic net
+    - Added Iterative interface with progress() method
+    - Iterative Learners renamed steps() method to progress()
+    - Changed default gradient-based `minChange` from 1e-4 to 1e-5
     
 - 2.5.15
     - Optimize random weighted subset with replacement